03 - DALLE-3 + API/requirements.txt
openai==1.57.3
python-dotenv==1.0.1
requests==2.25.1